requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altair=4.2.0=pypi_0
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
blinker=1.4=pypi_0
bottleneck=1.3.4=py38h67323c0_0
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py38h9ed2024_1003
ca-certificates=2022.4.26=hecd8cb5_0
cachetools=5.0.0=pypi_0
catalogue=2.0.7=py38hecd8cb5_0
certifi=2021.10.8=py38hecd8cb5_2
cffi=1.15.0=py38hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38hecd8cb5_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=36.0.0=py38hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
cymem=2.0.6=py38he9d5cce_0
cython-blis=0.7.7=py38h67323c0_0
debugpy=1.5.1=py38he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
en-core-web-md=3.3.0=pypi_0
entrypoints=0.4=py38hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
fastjsonschema=2.15.3=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
giflib=5.2.1=haf1e3a3_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
htbuilder=0.6.0=pypi_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.7.1=pypi_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.9.1=py38hecd8cb5_0
ipython=8.2.0=py38hecd8cb5_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
iteration-utilities=0.11.0=pypi_0
jedi=0.18.1=py38hecd8cb5_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
jsonschema=4.5.1=pypi_0
jupyter_client=7.2.2=py38hecd8cb5_0
jupyter_core=4.9.2=py38hecd8cb5_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.3.2=py38he9d5cce_0
langcodes=3.3.0=pyhd3eb1b0_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.2.0=h87d7836_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
markupsafe=2.0.1=py38h9ed2024_0
matplotlib=3.5.1=py38hecd8cb5_1
matplotlib-base=3.5.1=py38hfb0c5b7_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=pypi_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.1=py38h4ab4a9b_0
mkl_random=1.2.2=py38hb2f4e1b_0
munkres=1.1.4=py_0
murmurhash=1.0.7=py38he9d5cce_0
nbclient=0.6.2=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.5=py38hecd8cb5_0
notebook=6.4.11=pypi_0
numexpr=2.8.1=py38h2e5f0a9_0
numpy=1.21.5=py38h2e5f0a9_2
numpy-base=1.21.5=py38h3b1a694_2
openssl=1.1.1n=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py38he9d5cce_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pathy=0.6.1=py38hecd8cb5_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.1.0=pypi_0
pip=21.2.4=py38hecd8cb5_0
preshed=3.0.6=py38he9d5cce_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pydeck=0.7.1=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38_1
python=3.8.13=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyzmq=22.3.0=py38he9d5cce_2
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py38hae1ba45_1
scipy=1.7.3=py38h8c7af03_0
seaborn=0.11.2=pyhd3eb1b0_0
semver=2.13.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=61.2.0=py38hecd8cb5_0
shellingham=1.3.1=pyhd3eb1b0_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py38hecd8cb5_0
smmap=5.0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spacy=3.3.0=py38hc29d2bd_0
spacy-legacy=3.0.9=py38hecd8cb5_0
spacy-loggers=1.0.1=pyhd3eb1b0_0
sqlite=3.38.3=h707629a_0
srsly=2.4.3=py38he9d5cce_0
st-annotated-text=3.0.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.9.0=pypi_0
terminado=0.13.3=pypi_0
thinc=8.0.15=py38hc29d2bd_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h7bc2e8c_0
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
tornado=6.1=py38h9ed2024_0
tqdm=4.64.0=py38hecd8cb5_0
traitlets=5.1.1=pyhd3eb1b0_0
typer=0.4.1=py38hecd8cb5_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022.1=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.9=py38hecd8cb5_0
validators=0.19.0=pypi_0
wasabi=0.9.1=py38hecd8cb5_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
xz=5.2.5=hca72f7f_1
zeromq=4.3.4=h23ab428_0
zipp=3.8.0=pypi_0
zlib=1.2.12=h4dc903c_2
zstd=1.4.9=h322a384_0
